content/svg/content/test/test_transform.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=512636
-->
<head>
  <title>Test SVGTransform behavior</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=512636">Mozilla Bug 512636</a>
<p id="display"></p>
<div id="content">

  <svg xmlns="http://www.w3.org/2000/svg" width="100%" height="1" id="svg">
    <g id="g" transform="translate(10, 20)"/>
  </svg>

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

var tolerance = 1 / 65535;
      
function isequal( value, expected, tolerance )
{
  ok(Math.abs(value - expected) < tolerance, 'matrix value expected:' +expected + ' actual:' + value);
}

function run()
{
  var g, svg, t, m, m2;

  svg = $('svg');
  g = $('g');

  t = g.transform.baseVal.getItem(0);
  m = t.matrix;

  // test that the SVGTransform correctly reflects the translate()
  is(t.type, SVGTransform.SVG_TRANSFORM_TRANSLATE, 't.type for translate');
  is(m.a, 1, 'm.a for translate');
  is(m.b, 0, 'm.b for translate');
  is(m.c, 0, 'm.c for translate');
  is(m.d, 1, 'm.d for translate');
  is(m.e, 10, 'm.e for translate');
  is(m.f, 20, 'm.f for translate');
  is(t.angle, 0, 't.angle for translate');

  // set the SVGTransform to be a scale()
  t.setScale(2, 3);

  // test that the matrix is live and now reflects the scale()
  is(t.type, SVGTransform.SVG_TRANSFORM_SCALE, 't.type for scale');
  is(m.a, 2, 'm.a for scale');
  is(m.b, 0, 'm.b for scale');
  is(m.c, 0, 'm.c for scale');
  is(m.d, 3, 'm.d for scale');
  is(m.e, 0, 'm.e for scale');
  is(m.f, 0, 'm.f for scale');
  is(t.angle, 0, 't.angle for scale');

  // set the SVGTransform to be a matrix()
  m2 = svg.createSVGMatrix();
  m2.a = 1;
  m2.b = 2;
  m2.c = 3;
  m2.d = 4;
  m2.e = 5;
  m2.f = 6;
  t.setMatrix(m2);

  // check that setMatrix() took a copy of m
  ok(m != m2, 't.matrix identity');

  // test that the SVGTransform now reflects the matrix value
  is(t.type, SVGTransform.SVG_TRANSFORM_MATRIX, 't.type for matrix');
  is(m.a, 1, 'm.a for matrix');
  is(m.b, 2, 'm.b for matrix');
  is(m.c, 3, 'm.c for matrix');
  is(m.d, 4, 'm.d for matrix');
  is(m.e, 5, 'm.e for matrix');
  is(m.f, 6, 'm.f for matrix');
  is(t.angle, 0, 't.angle for matrix');

  // set the SVGTransform to be a translate() then convert to a matrix
  t.setTranslate(0, 0);
  m.a = 2;

  // test that the SVGTransform now reflects the matrix value
  is(t.type, SVGTransform.SVG_TRANSFORM_MATRIX, 't.type for matrix');

  // set the SVGTransform to be a rotate()
  t.setRotate(90, 0, 0);

  // test that the SVGTransform now reflects the matrix value
  is(t.type, SVGTransform.SVG_TRANSFORM_ROTATE, 't.type for rotate');
  isequal(m.a, Math.cos(Math.PI/2), tolerance);
  isequal(m.b, Math.sin(Math.PI/2), tolerance);
  isequal(m.c, -Math.sin(Math.PI/2), tolerance);
  isequal(m.d, Math.cos(Math.PI/2), tolerance);
  isequal(m.e, 0, tolerance);
  isequal(m.f, 0, tolerance);

  // set the SVGTransform to be a skewX()
  t.setSkewX(45);

  // test that the SVGTransform now reflects the matrix value
  is(t.type, SVGTransform.SVG_TRANSFORM_SKEWX, 't.type for skewx');
  isequal(m.a, 1, tolerance);
  isequal(m.b, 0, tolerance);
  isequal(m.c, Math.tan(Math.PI/4), tolerance);
  isequal(m.d, Math.tan(Math.PI/4), tolerance);
  isequal(m.e, 0, tolerance);
  isequal(m.f, 0, tolerance);

  // set the SVGTransform to be a skewY()
  t.setSkewY(45);

  // test that the SVGTransform now reflects the matrix value
  is(t.type, SVGTransform.SVG_TRANSFORM_SKEWY, 't.type for skewy');
  isequal(m.a, Math.tan(Math.PI/4), tolerance);
  isequal(m.b, Math.tan(Math.PI/4), tolerance);
  isequal(m.c, 0, tolerance);
  isequal(m.d, 1, tolerance);
  isequal(m.e, 0, tolerance);
  isequal(m.f, 0, tolerance);

  SimpleTest.finish();
}

window.addEventListener("load", run, false);

]]>
</script>
</pre>
</body>
</html>
